requirements.txt
bitarray==2.9.2
bitstring==4.2.3
cffi==1.17.0
contourpy==1.3.1
cryptography==43.0.0
cycler==0.12.1
distlib==0.3.7
ecdsa==0.19.0
esptool==4.7.0
filelock==3.13.1
fonttools==4.55.2
intelhex==2.3.0
kiwisolver==1.4.7
matplotlib==3.9.3
numpy==2.1.3
packaging==24.2
pbr==5.11.1
pillow==11.0.0
platformdirs==3.11.0
pycparser==2.22
pyparsing==3.2.0
pyserial==3.5
python-dateutil==2.9.0.post0
PyYAML==6.0.2
reedsolo==1.7.0
six==1.16.0
stevedore==5.1.0
virtualenv==20.24.6
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
